# docs/api/swagger.yaml
openapi: 3.0.3
info:
  title: Personal Finance API - Gastos
  description: API para gestión de gastos en aplicación de finanzas personales
  version: 1.0.0
  contact:
    name: Personal Finance API
    url: http://localhost:3030
servers:
  - url: http://localhost:3030/api
    description: Development server
...
paths:
...
  /gastos-recurrentes:
    get:
      tags:
        - gastos-recurrentes
      summary: Obtener gastos recurrentes con filtros
      description: |
        Obtiene gastos recurrentes con filtros específicos para frecuencia y estado.
      parameters:
        - name: categoria_gasto_id
          in: query
          schema:
            type: integer
        - name: importancia_gasto_id
          in: query
          schema:
            type: integer
        - name: tipo_pago_id
          in: query
          schema:
            type: integer
        - name: tarjeta_id
          in: query
          schema:
            type: integer
        - name: frecuencia_gasto_id
          in: query
          schema:
            type: integer
          description: Filtrar por frecuencia (mensual, anual, etc.)
        - name: monto_min
          in: query
          schema:
            type: number
        - name: monto_max
          in: query
          schema:
            type: number
        - name: activo
          in: query
          schema:
            type: boolean
          description: Filtrar por estado activo/inactivo
        - name: dia_de_pago
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 31
          description: Día específico del mes
        - name: mes_de_pago
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 12
          description: Mes específico (para gastos anuales)
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: orderBy
          in: query
          schema:
            type: string
            enum: [monto, descripcion, dia_de_pago, mes_de_pago, activo, createdAt, updatedAt]
            default: createdAt
        - name: orderDirection
          in: query
          schema:
            type: string
            enum: [ASC, DESC]
            default: DESC
      responses:
        '200':
          description: Lista de gastos recurrentes

    post:
      tags:
        - gastos-recurrentes
      summary: Crear gasto recurrente
      description: |
        Crea un nuevo gasto recurrente. Se procesará automáticamente según su frecuencia.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - descripcion
                - monto
                - dia_de_pago
                - categoria_gasto_id
                - importancia_gasto_id
                - tipo_pago_id
                - frecuencia_gasto_id
              properties:
                descripcion:
                  type: string
                  minLength: 3
                  maxLength: 255
                monto:
                  type: number
                  minimum: 0.01
                dia_de_pago:
                  type: integer
                  minimum: 1
                  maximum: 31
                mes_de_pago:
                  type: integer
                  minimum: 1
                  maximum: 12
                  nullable: true
                  description: Requerido para frecuencias anuales
                categoria_gasto_id:
                  type: integer
                  minimum: 1
                importancia_gasto_id:
                  type: integer
                  minimum: 1
                tipo_pago_id:
                  type: integer
                  minimum: 1
                frecuencia_gasto_id:
                  type: integer
                  minimum: 1
                tarjeta_id:
                  type: integer
                  minimum: 1
                  nullable: true
                fecha_inicio:
                  type: string
                  format: date
                  nullable: true
      responses:
        '201':
          description: Gasto recurrente creado exitosamente
...